My update2

+++ Server/MirObjects/Monsters/CharmedSnake.cs
@@ -255,7 +255,7 @@ namespace Server.MirObjects.Monsters
                 Image = Monster.CharmedSnake,
                 Direction = Direction,
                 Effect = Info.Effect,
-                AI = Info.AI,
+                AI = (byte)Info.AI,
                 Light = Info.Light,
                 Dead = Dead,
                 Skeleton = Harvested,
