+++ Server/MirObjects/Monsters/HolyDeva.cs
@@ -114,7 +114,7 @@ namespace Server.MirObjects.Monsters
                 Image = Monster.HolyDeva,
                 Direction = Direction,
                 Effect = Info.Effect,
-                AI = Info.AI,
+                AI = (byte)Info.AI,
                 Light = Info.Light,
                 Dead = Dead,
                 Skeleton = Harvested,
