+++ Server/MirObjects/Monsters/HornedMage.cs
@@ -124,7 +124,7 @@ namespace Server.MirObjects.Monsters
                 Image = Info.Image,
                 Direction = Direction,
                 Effect = Info.Effect,
-                AI = Info.AI,
+                AI = (byte)Info.AI,
                 Light = Info.Light,
                 Dead = Dead,
                 Skeleton = Harvested,
