+++ Server/MirObjects/Monsters/SpittingToad.cs
@@ -143,7 +143,7 @@ namespace Server.MirObjects.Monsters
                 Image = Monster.SpittingToad,
                 Direction = Direction,
                 Effect = Info.Effect,
-                AI = Info.AI,
+                AI = (byte)Info.AI,
                 Light = Info.Light,
                 Dead = Dead,
                 Skeleton = Harvested,
