+++ Server/MirObjects/Monsters/VampireSpider.cs
@@ -202,7 +202,7 @@ namespace Server.MirObjects.Monsters
                     Image = Monster.VampireSpider,
                     Direction = Direction,
                     Effect = Info.Effect,
-                    AI = Info.AI,
+                    AI = (byte)Info.AI,
                     Light = Info.Light,
                     Dead = Dead,
                     Skeleton = Harvested,
